Implement Queryable + Updatable across all cache-based sources
Provide a consistent implementation of the `Queryable` and `Updatable` interfaces across all record-cache-based sources:

* `MemorySource`
* `LocalStorageSource`
* `IndexedDBSource`

This will allow for the deprecation of `Pullable` and `Pushable`, now that equivalent results can be achieved via `query` and `update` calls with the `fullResponse: true` option. Reducing the number of interfaces per-source should be a nice simplification that allows for more flexibility.
