+++ packages/@orbit/indexeddb/src/indexeddb-source.ts
@@ -1,29 +1,37 @@
 import { Orbit } from '@orbit/core';
 import {
   buildTransform,
+  DefaultRequestOptions,
+  FullResponse,
   pullable,
   pushable,
+  queryable,
+  RequestOptions,
   Resettable,
+  ResponseHints,
   syncable,
-  FullResponse,
-  DefaultRequestOptions,
-  RequestOptions
+  updatable
 } from '@orbit/data';
 import {
   RecordOperation,
-  UpdateRecordOperation,
-  RecordQueryExpressionResult,
-  RecordSourceSettings,
+  RecordOperationResult,
   RecordPullable,
   RecordPushable,
+  RecordQuery,
+  RecordQueryable,
+  RecordQueryExpressionResult,
+  RecordQueryResult,
+  RecordSource,
+  RecordSourceQueryOptions,
+  RecordSourceSettings,
   RecordSyncable,
   RecordTransform,
-  RecordSource,
-  RecordQuery,
-  RecordSourceQueryOptions
+  RecordTransformResult,
+  RecordUpdatable,
+  UpdateRecordOperation
 } from '@orbit/records';
-import { supportsIndexedDB } from './lib/indexeddb';
 import { IndexedDBCache, IndexedDBCacheSettings } from './indexeddb-cache';
+import { supportsIndexedDB } from './lib/indexeddb';
 
 const { assert } = Orbit;
 
@@ -37,6 +45,8 @@ export interface IndexedDBSource
     RecordSyncable,
     RecordPullable<unknown>,
     RecordPushable<unknown>,
+    RecordQueryable<unknown>,
+    RecordUpdatable<unknown>,
     Resettable {}
 
 /**
@@ -44,6 +54,8 @@ export interface IndexedDBSource
  */
 @pullable
 @pushable
+@queryable
+@updatable
 @syncable
 export class IndexedDBSource extends RecordSource {
   protected _cache: IndexedDBCache;
@@ -137,9 +149,92 @@ export class IndexedDBSource extends RecordSource {
 
   async _sync(transform: RecordTransform): Promise<void> {
     if (!this.transformLog.contains(transform.id)) {
-      await this._cache.update(transform);
+      await this._applyTransform(transform);
       await this.transformed([transform]);
     }
+  }
+
+  /////////////////////////////////////////////////////////////////////////////
+  // Updatable interface implementation
+  /////////////////////////////////////////////////////////////////////////////
+
+  async _update(
+    transform: RecordTransform,
+    hints?: ResponseHints<RecordTransformResult, unknown>
+  ): Promise<FullResponse<RecordTransformResult, unknown, RecordOperation>> {
+    let results: RecordTransformResult;
+    const response: FullResponse<
+      RecordTransformResult,
+      unknown,
+      RecordOperation
+    > = {};
+
+    if (!this.transformLog.contains(transform.id)) {
+      results = await this._applyTransform(transform);
+      response.transforms = [transform];
+    }
+
+    if (hints?.data) {
+      if (transform.operations.length > 1 && Array.isArray(hints.data)) {
+        const responseData = [];
+        const hintsData = hints.data as RecordOperationResult[];
+        for (let h of hintsData) {
+          responseData.push(await this._retrieveOperationResult(h));
+        }
+        response.data = responseData;
+      } else {
+        response.data = await this._retrieveOperationResult(
+          hints.data as RecordOperationResult
+        );
+      }
+    } else if (results) {
+      if (transform.operations.length === 1 && Array.isArray(results)) {
+        response.data = results[0];
+      } else {
+        response.data = results;
+      }
+    }
+
+    if (hints?.details) {
+      response.details = hints.details;
+    }
+
+    return response;
+  }
+
+  /////////////////////////////////////////////////////////////////////////////
+  // Queryable interface implementation
+  /////////////////////////////////////////////////////////////////////////////
+
+  async _query(
+    query: RecordQuery,
+    hints?: ResponseHints<RecordQueryResult, unknown>
+  ): Promise<FullResponse<RecordQueryResult, unknown, RecordOperation>> {
+    let response: FullResponse<RecordQueryResult, unknown, RecordOperation>;
+
+    if (hints?.data) {
+      response = {};
+      if (query.expressions.length > 1 && Array.isArray(hints.data)) {
+        const responseData = [];
+        const hintsData = hints.data as RecordQueryExpressionResult[];
+        for (let h of hintsData) {
+          responseData.push(await this._retrieveQueryExpressionResult(h));
+        }
+        response.data = responseData;
+      } else {
+        response.data = await this._retrieveQueryExpressionResult(
+          hints.data as RecordQueryExpressionResult
+        );
+      }
+    } else {
+      response = await this._cache.query(query, { fullResponse: true });
+    }
+
+    if (hints?.details) {
+      response.details = hints.details;
+    }
+
+    return response;
   }
 
   /////////////////////////////////////////////////////////////////////////////
@@ -185,6 +280,38 @@ export class IndexedDBSource extends RecordSource {
     fullResponse.transforms = [buildTransform(operations)];
 
     return fullResponse;
+  }
+
+  /////////////////////////////////////////////////////////////////////////////
+  // Protected methods
+  /////////////////////////////////////////////////////////////////////////////
+
+  protected async _retrieveQueryExpressionResult(
+    result: RecordQueryExpressionResult
+  ): Promise<RecordQueryExpressionResult> {
+    if (Array.isArray(result)) {
+      return this._cache.getRecordsAsync(result);
+    } else if (result) {
+      return this._cache.getRecordAsync(result);
+    } else {
+      return result;
+    }
+  }
+
+  protected async _retrieveOperationResult(
+    result: RecordOperationResult
+  ): Promise<RecordOperationResult> {
+    if (result) {
+      return this._cache.getRecordAsync(result);
+    } else {
+      return result;
+    }
+  }
+
+  protected async _applyTransform(
+    transform: RecordTransform
+  ): Promise<RecordTransformResult> {
+    return await this.cache.update(transform);
   }
 
   protected _operationsFromQueryResult(
